+++ py/fastspecfit/fastspecfit.py
@@ -40,7 +40,8 @@ def _assign_units_to_columns(fastfit, metadata, Spec, templates, fastphot, stack
 
 def fastspec_one(iobj, data, out, meta, templates, log=None,
                  minspecwave=3500., maxspecwave=9900., broadlinefit=True,
-                 fastphot=False, stackfit=False, nophoto=False, percamera_models=False):
+                 fastphot=False, stackfit=False, nophoto=False,
+                 percamera_models=False, debug_plots=False):
     """Multiprocessing wrapper to run :func:`fastspec` on a single object.
 
     """
@@ -60,7 +61,7 @@ def fastspec_one(iobj, data, out, meta, templates, log=None,
 
     continuummodel, smooth_continuum = continuum_specfit(data, out, templatecache,
                                                          fastphot=fastphot, nophoto=nophoto,
-                                                         log=log)
+                                                         debug_plots=debug_plots, log=log)
 
     # Optionally fit the emission-line spectrum.
     if fastphot:
@@ -125,6 +126,7 @@ def parse(options=None, log=None):
     parser.add_argument('--mapdir', type=str, default=None, help='Optional directory name for the dust maps.')
     parser.add_argument('--dr9dir', type=str, default=None, help='Optional directory name for the DR9 photometry.')
     parser.add_argument('--specproddir', type=str, default=None, help='Optional directory name for the spectroscopic production.')
+    parser.add_argument('--debug-plots', action='store_true', help='Generate a variety of debugging plots (written to $PWD).')
     parser.add_argument('--verbose', action='store_true', help='Be verbose (for debugging purposes).')
 
     if log is None:
@@ -225,7 +227,7 @@ def fastspec(fastphot=False, stackfit=False, args=None, comm=None, verbose=False
     t0 = time.time()
     fitargs = [(iobj, data[iobj], out[iobj], meta[iobj], templates, log,
                 minspecwave, maxspecwave, args.broadlinefit, fastphot, stackfit,
-                args.nophoto, args.percamera_models)
+                args.nophoto, args.percamera_models, args.debug_plots)
                 for iobj in np.arange(Spec.ntargets)]
     if args.mp > 1:
         import multiprocessing
@@ -640,10 +642,10 @@ def qa_fastspec(data, templatecache, fastspec, metadata, coadd_type='healpix',
          
         for oneline in EMFit.linetable[inrange]: # for all lines in range
             linename = oneline['name'].upper()
-            amp = fastspec['{}_AMP'.format(linename)]
+            amp = fastspec['{}_MODELAMP'.format(linename)]
             if amp != 0:
                 desiemlines_oneline1 = build_emline_model(
-                    EMFit.log10wave, redshift, np.array([amp]),
+                    EMFit.dlog10wave, redshift, np.array([amp]),
                     np.array([fastspec['{}_VSHIFT'.format(linename)]]),
                     np.array([fastspec['{}_SIGMA'.format(linename)]]),
                     np.array([oneline['restwave']]), data['wave'], data['res'])
@@ -1016,11 +1018,13 @@ def qa_fastspec(data, templatecache, fastspec, metadata, coadd_type='healpix',
         plotsig_default_balmer = 500.0 # [km/s]
         plotsig_default_broad = 2000.0 # [km/s]
     
-        meanwaves, deltawaves, sigmas, linenames = [], [], [], []
+        minwaves, maxwaves, meanwaves, deltawaves, sigmas, linenames = [], [], [], [], [], []
         for plotgroup in set(linetable['plotgroup']):
             I = np.where(plotgroup == linetable['plotgroup'])[0]
             linename = linetable['nicename'][I[0]].replace('-', ' ')
             linenames.append(linename)
+            minwaves.append(np.min(linetable['restwave'][I]))
+            maxwaves.append(np.max(linetable['restwave'][I]))
             meanwaves.append(np.mean(linetable['restwave'][I]))
             deltawaves.append((np.max(linetable['restwave'][I]) - np.min(linetable['restwave'][I])) / 2)
         
@@ -1053,6 +1057,8 @@ def qa_fastspec(data, templatecache, fastspec, metadata, coadd_type='healpix',
             ax = []
         else:
             srt = np.argsort(meanwaves)
+            minwaves = np.hstack(minwaves)[srt]
+            maxwaves = np.hstack(maxwaves)[srt]
             meanwaves = np.hstack(meanwaves)[srt]
             deltawaves = np.hstack(deltawaves)[srt]
             sigmas = np.hstack(sigmas)[srt]
@@ -1083,7 +1089,8 @@ def qa_fastspec(data, templatecache, fastspec, metadata, coadd_type='healpix',
             else:
                 ax, irow, colshift = [], 4, 5 # skip the gap row
                 
-            for iax, (meanwave, deltawave, sig, linename) in enumerate(zip(meanwaves, deltawaves, sigmas, linenames)):
+            for iax, (minwave, maxwave, meanwave, deltawave, sig, linename) in enumerate(
+                    zip(minwaves, maxwaves, meanwaves, deltawaves, sigmas, linenames)):
                 icol = iax % nlinecols
                 icol += colshift
                 if iax > 0 and iax % nlinecols == 0:
@@ -1093,9 +1100,10 @@ def qa_fastspec(data, templatecache, fastspec, metadata, coadd_type='healpix',
                 xx = fig.add_subplot(gs[irow, icol])
                 ax.append(xx)
             
-                wmin = (meanwave - deltawave) * (1+redshift) - 6 * sig * meanwave * (1+redshift) / C_LIGHT
-                wmax = (meanwave + deltawave) * (1+redshift) + 6 * sig * meanwave * (1+redshift) / C_LIGHT
-                #print(linename, wmin, wmax)
+                wmin = (minwave - deltawave) * (1+redshift) - 6 * sig * minwave * (1+redshift) / C_LIGHT
+                wmax = (maxwave + deltawave) * (1+redshift) + 6 * sig * maxwave * (1+redshift) / C_LIGHT
+                #wmin = (meanwave - deltawave) * (1+redshift) - 6 * sig * meanwave * (1+redshift) / C_LIGHT
+                #wmax = (meanwave + deltawave) * (1+redshift) + 6 * sig * meanwave * (1+redshift) / C_LIGHT
             
                 # iterate over cameras
                 for icam in np.arange(len(data['cameras'])): # iterate over cameras
@@ -1109,9 +1117,6 @@ def qa_fastspec(data, templatecache, fastspec, metadata, coadd_type='healpix',
                     emlinesigma = emlinesigma[good]
                     emlinemodel = emlinemodel[good]
             
-                    #if icam == 0:
-                    #    import matplotlib.pyplot as plt ; plt.clf() ; plt.plot(emlinewave, emlineflux) ; plt.plot(emlinewave, emlinemodel) ; plt.xlim(4180, 4210) ; plt.ylim(-15, 17) ; plt.savefig('desi-users/ioannis/tmp/junkg.png')
-                        
                     emlinemodel_oneline = []
                     for desiemlines_oneline1 in desiemlines_oneline:
                         emlinemodel_oneline.append(desiemlines_oneline1[icam][good])